Move ReadOffsetTable to postings.go.
