+++ pkg/storegateway/indexheader/index/index.go
@@ -7,7 +7,6 @@ package index
 
 import (
 	"hash/crc32"
-	"os"
 	"sort"
 	"unsafe"
 
@@ -155,36 +154,6 @@ func (s *Symbols) ReverseLookup(sym string) (uint32, error) {
 
 func (s *Symbols) Size() int {
 	return len(s.offsets) * 8
-}
-
-// ReadOffsetTable reads an offset table and at the given position calls f for each
-// found entry. If f returns an error it stops decoding and returns the received error.
-func ReadOffsetTable(src *os.File, off uint64, f func([]string, uint64, int) error) error {
-	d := stream_encoding.NewDecbufFromFile(src, int(off), castagnoliTable)
-	startLen := d.Len()
-	cnt := d.Be32()
-
-	for d.Err() == nil && d.Len() > 0 && cnt > 0 {
-		offsetPos := startLen - d.Len()
-		keyCount := d.Uvarint()
-		// The Postings offset table takes only 2 keys per entry (name and value of label),
-		// and the LabelIndices offset table takes only 1 key per entry (a label name).
-		// Hence setting the size to max of both, i.e. 2.
-		keys := make([]string, 0, 2)
-
-		for i := 0; i < keyCount; i++ {
-			keys = append(keys, d.UvarintStr())
-		}
-		o := d.Uvarint64()
-		if d.Err() != nil {
-			break
-		}
-		if err := f(keys, o, offsetPos); err != nil {
-			return err
-		}
-		cnt--
-	}
-	return d.Err()
 }
 
 func yoloString(b []byte) string {
