+++ pkg/storegateway/indexheader/index/postings.go
@@ -6,7 +6,11 @@
 package index
 
 import (
+	"os"
+
 	"github.com/prometheus/prometheus/model/labels"
+
+	stream_encoding "github.com/grafana/mimir/pkg/storegateway/indexheader/encoding"
 )
 
 var allPostingsKey = labels.Label{}
@@ -14,4 +18,34 @@ var allPostingsKey = labels.Label{}
 // AllPostingsKey returns the label key that is used to store the postings list of all existing IDs.
 func AllPostingsKey() (name, value string) {
 	return allPostingsKey.Name, allPostingsKey.Value
+}
+
+// ReadOffsetTable reads an offset table and at the given position calls f for each
+// found entry. If f returns an error it stops decoding and returns the received error.
+func ReadOffsetTable(src *os.File, off uint64, f func([]string, uint64, int) error) error {
+	d := stream_encoding.NewDecbufFromFile(src, int(off), castagnoliTable)
+	startLen := d.Len()
+	cnt := d.Be32()
+
+	for d.Err() == nil && d.Len() > 0 && cnt > 0 {
+		offsetPos := startLen - d.Len()
+		keyCount := d.Uvarint()
+		// The Postings offset table takes only 2 keys per entry (name and value of label),
+		// and the LabelIndices offset table takes only 1 key per entry (a label name).
+		// Hence setting the size to max of both, i.e. 2.
+		keys := make([]string, 0, 2)
+
+		for i := 0; i < keyCount; i++ {
+			keys = append(keys, d.UvarintStr())
+		}
+		o := d.Uvarint64()
+		if d.Err() != nil {
+			break
+		}
+		if err := f(keys, o, offsetPos); err != nil {
+			return err
+		}
+		cnt--
+	}
+	return d.Err()
 }
